Fix "configuration run in progress" failure
In GHE 2.14, you can no longer disable maintenance mode while a
configuration run is in progress. Running the ghe-maintenance
command in this state results in a failed exit code, which fails
a playbook run.

This commit handles this case more gracefully by retrying the
command to disable maintenance mode until it completes with an exit
code of 0.